fix : 웨딩관련 type을 파일위치 이동
