+++ apps/client/src/features/wedding/types.ts
@@ -1,5 +1,11 @@
-export interface WeddingInfo {
+export type WeddingInfoType = {
   weddingDate: string;
   weddingTime: string;
   weddingHall: string;
-}
+};
+
+export type WeddingBudgetInfoType = {
+  totalBudget: number;
+  totalSpending: number;
+  spendingPerBudget: number;
+};
