+++ apps/client/src/shared/apis/wedding.ts
@@ -1,11 +1,13 @@
 import { axiosInstance } from '../config/axios';
-import { WeddingBudgetInfoType, WeddingInfoType } from '../types/wedding';
+import { WeddingBudgetInfoType, WeddingInfoType } from '@fsd/features/wedding/types';
 
 const URL_ROOT = '/wedding';
 
 export const WeddingClient = {
-  getWeddingInfo: async (): Promise<WeddingInfoType | WeddingBudgetInfoType> => {
-    return await axiosInstance.get(URL_ROOT);
+  getWeddingInfo: async (): Promise<WeddingInfoType & WeddingBudgetInfoType> => {
+    const { data } = await axiosInstance.get(URL_ROOT);
+
+    return data;
   },
   createWeddingInfo: async (formData: WeddingInfoType) => {
     return await axiosInstance.post(URL_ROOT, { ...formData });
@@ -14,7 +16,7 @@ export const WeddingClient = {
     return await axiosInstance.put(URL_ROOT, { ...formData });
   },
   updateTotalBudget: async (totalBudget: number) => {
-    await axiosInstance.put('/totalBudget', {
+    await axiosInstance.put(URL_ROOT + '/total-budget', {
       totalBudget,
     });
   },
