--- apps/client/src/shared/types/wedding.ts
@@ -1,11 +1,0 @@
-export type WeddingInfoType = {
-  weddingDate: string;
-  weddingTime: string;
-  weddingHall: string;
-};
-
-export type WeddingBudgetInfoType = {
-  totalBudget: number;
-  totalSpending: number;
-  spendingPerBudget: number;
-};
